Migrate JAX-RS Consumes and Produces annotations on class level
