+++ components/sbm-recipes-jee-to-boot/src/test/java/org/springframework/sbm/jee/jaxrs/recipes/ConvertJaxRsAnnotationsTest.java
@@ -13,13 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.springframework.sbm.jee.jaxrs.actions;
+package org.springframework.sbm.jee.jaxrs.recipes;
 
 import org.intellij.lang.annotations.Language;
 import org.junit.jupiter.api.Test;
 import org.springframework.sbm.engine.context.ProjectContext;
-import org.springframework.sbm.java.api.Method;
 import org.springframework.sbm.java.migration.conditions.HasTypeAnnotation;
+import org.springframework.sbm.jee.jaxrs.actions.ConvertJaxRsAnnotations;
 import org.springframework.sbm.project.resource.TestProjectContext;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -27,12 +27,12 @@ import static org.assertj.core.api.Assertions.assertThat;
 /**
  * @author Fabian Krüger
  */
-public class MigrateJaxRsAnnotations_Test {
+public class ConvertJaxRsAnnotationsTest {
 
     private final static String SPRING_VERSION = "5.3.13";
 
     @Test
-    void convertJaxRsMethodToSpringMvc() {
+    void noPathOnMethodLevel() {
         @Language("java")
         String restControllerCode = """
                 package com.example.jeerest.rest;
@@ -54,6 +54,13 @@ public class MigrateJaxRsAnnotations_Test {
                 @Path("movies")
                 @Produces({"application/json"})
                 public class MoviesRest {
+                                
+                    @GET
+                    @Path("{id}")
+                    public Movie find(@PathParam("id") Long id) {
+                        return service.find(id);
+                    }
+                                
                     @GET
                     public List<Movie> getMovies(@QueryParam("first") Integer first, @QueryParam("max") Integer max,
                                                  @QueryParam("field") String field, @QueryParam("searchTerm") String searchTerm) {
@@ -62,20 +69,14 @@ public class MigrateJaxRsAnnotations_Test {
                 }
                 """;
 
-        ProjectContext context = TestProjectContext
-                .buildProjectContext()
+        ProjectContext context = TestProjectContext.buildProjectContext()
                 .withJavaSources(restControllerCode)
-                .withBuildFileHavingDependencies("org.jboss.spec.javax.ws.rs:jboss-jaxrs-api_2.1_spec:1.0.1.Final",
-                                                 "org.springframework:spring-core:" + SPRING_VERSION)
+                .withBuildFileHavingDependencies(
+                        "org.jboss.spec.javax.ws.rs:jboss-jaxrs-api_2.1_spec:1.0.1.Final",
+                        "org.springframework:spring-core:"+SPRING_VERSION,
+                        "org.springframework:spring-web:"+SPRING_VERSION
+                )
                 .build();
-
-        Method jaxRsMethod =  context.getProjectJavaSources()
-                .list()
-                .get(0)
-                .getTypes()
-                .get(0)
-                .getMethods()
-                .get(0);
 
         ConvertJaxRsAnnotations convertJaxRsAnnotations = ConvertJaxRsAnnotations
                 .builder()
@@ -83,7 +84,7 @@ public class MigrateJaxRsAnnotations_Test {
                 .description("Convert JAX-RS annotations into Spring Boot annotations.")
                 .build();
 
-        convertJaxRsAnnotations.convertJaxRsMethodToSpringMvc(jaxRsMethod);
+        convertJaxRsAnnotations.apply(context);
 
 
         @Language("java")
@@ -94,19 +95,27 @@ public class MigrateJaxRsAnnotations_Test {
                 import com.example.jeerest.Movie;
                 import com.example.jeerest.MoviesBean;
                 import org.springframework.beans.factory.annotation.Autowired;
+                import org.springframework.web.bind.annotation.RequestMapping;
+                import org.springframework.web.bind.annotation.RequestMethod;
+                import org.springframework.web.bind.annotation.RestController;
                 
                 import javax.ws.rs.DELETE;
                 import javax.ws.rs.PUT;
-                import javax.ws.rs.Path;
                 import javax.ws.rs.PathParam;
-                import javax.ws.rs.Produces;
                 import javax.ws.rs.QueryParam;
                 import javax.ws.rs.core.MediaType;
                 import java.util.List;
                 
-                @Path("movies")
-                @Produces({"application/json"})
+                
+                @RestController
+                @RequestMapping(value = "movies", produces = {"application/json"})
                 public class MoviesRest {
+                
+                    @RequestMapping(value = "{id}", method = RequestMethod.GET)
+                    public Movie find(@PathParam("id") Long id) {
+                        return service.find(id);
+                    }
+                
                     @RequestMapping(method = RequestMethod.GET)
                     public List<Movie> getMovies(@QueryParam("first") Integer first, @QueryParam("max") Integer max,
                                                  @QueryParam("field") String field, @QueryParam("searchTerm") String searchTerm) {
